GEODE-3618: Move GfshRule into geode-junit
