+++ geode-junit/build.gradle
@@ -25,7 +25,8 @@ dependencies {
   compile 'com.google.guava:guava:' + project.'guava.version'
   compile 'org.mockito:mockito-core:' + project.'mockito-core.version'
   compile 'org.awaitility:awaitility:' + project.'awaitility.version'
-
+  compile 'org.apache.logging.log4j:log4j-api:' + project.'log4j.version'
+  compile 'org.apache.logging.log4j:log4j-core:' + project.'log4j.version'
 
   compile('junit:junit:' + project.'junit.version') {
     exclude module: 'hamcrest-core'
